Ôö¼ÓÕæ¼Ùɾ³ý

+++ del.c
@@ -10,6 +10,7 @@ int cgiMain()
 	fprintf(cgiOut, "Content-type:text/html;charset=utf-8\n\n");
 
 	char stuId[32] = "\0";
+	char flag[32] = "\0";
 	int status = 0;
 
 
@@ -17,6 +18,12 @@ int cgiMain()
 	if (status != cgiFormSuccess)
 	{
 		fprintf(cgiOut, "get stuId error!\n");
+		return 1;
+	}
+	status = cgiFormString("flag",  flag, 32);
+	if (status != cgiFormSuccess)
+	{
+		fprintf(cgiOut, "get flag error!\n");
 		return 1;
 	}
 
@@ -42,13 +49,22 @@ int cgiMain()
 		return -1;
 	}
 
-
-	sprintf(sql, "delete from stu where stuId = %d", atoi(stuId));
-	if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
-	{
+	if (flag[0] == '1'){
+		sprintf(sql, "delete from stu where stuId = %d", atoi(stuId));
+		if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
+		{
 		fprintf(cgiOut,"mysql_real_query fail:%s\n", mysql_error(db));
 		mysql_close(db);
 		return -1;
+		}
+	}else{
+		sprintf(sql, "update stu set state = '0' where stuId= %d", atoi(stuId));
+		if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
+		{
+		fprintf(cgiOut,"mysql_real_query fail:%s\n", mysql_error(db));
+		mysql_close(db);
+		return -1;
+		}
 	}
